requirements.txt
aiohttp>=3.9.0
websockets>=12.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
sse-starlette>=1.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
